pyproject.toml
[project]
name = "repl-nix-studyprotocolcreator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.52.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "streamlit>=1.39.0",
]
